# swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Specifikace REST API služby EET Gateway.",
        "title": "EET Gateway API",
        "contact": {
            "name": "Tommy Chu",
            "email": "pro@chutommy.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/certs": {
            "post": {
                "description": "Nahraje certifikát do služby keystore a umožní příslušnému daňovému poplatníkovi odesílat e-tržby v systému EET Gateway.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certifikáty"
                ],
                "summary": "Nahrát certifikát daňového poplatníka",
                "operationId": "storeCert",
                "parameters": [
                    {
                        "description": "Tělo požadavku na nahrání certifikátu.",
                        "name": "cert_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StoreCertificateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dotaz byl úspěšně zpracovaný.",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Vložený certifikát daňového poplatníka je neplatný.",
                        "schema": {
                            "$ref": "#/definitions/GatewayErrorResponse"
                        }
                    },
                    "409": {
                        "description": "ID certifikátu je již zabraný. Konfliktní certifikát musí být odstraněn nebo pro tento zvoleno jiné ID.",
                        "schema": {
                            "$ref": "#/definitions/GatewayErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Naskytla se neočekávaná chyba databáze.",
                        "schema": {
                            "$ref": "#/definitions/GatewayErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Služba keystore je nedostupná.",
                        "schema": {
                            "$ref": "#/definitions/GatewayErrorResponse"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Ověří stav API služby EET Gateway, keystore (službu spravující databázi) a spojení se servery správce daně.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Ověřit stav EETG API",
                "operationId": "ping",
                "responses": {
                    "200": {
                        "description": "Všechny EETG komponenty jsou dostupné.",
                        "schema": {
                            "$ref": "#/definitions/PingResponse"
                        }
                    },
                    "503": {
                        "description": "Některé EETG komponeny jsou nedostupné.",
                        "schema": {
                            "$ref": "#/definitions/PingResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "GatewayErrorResponse": {
            "type": "object",
            "properties": {
                "gateway_error": {
                    "type": "string",
                    "example": "keystore service unavailable"
                }
            }
        },
        "PingResponse": {
            "type": "object",
            "properties": {
                "eet_gateway": {
                    "type": "string",
                    "example": "online"
                },
                "keystore": {
                    "type": "string",
                    "example": "online"
                },
                "tax_admin": {
                    "type": "string",
                    "example": "online"
                }
            }
        },
        "StoreCertificateRequest": {
            "type": "object",
            "required": [
                "cert_id",
                "cert_password",
                "pkcs12_data",
                "pkcs12_password"
            ],
            "properties": {
                "cert_id": {
                    "type": "string",
                    "example": "d406ccda-1bc5-44ab-a081-af6e8740634c"
                },
                "cert_password": {
                    "type": "string",
                    "example": "10f47cc7eb3c51"
                },
                "pkcs12_data": {
                    "type": "string",
                    "format": "base64",
                    "example": "MIIPoq7Huhj ... HXJBdPA="
                },
                "pkcs12_password": {
                    "type": "string",
                    "example": "1234"
                }
            }
        },
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "cert_id": {
                    "type": "string",
                    "example": "d406ccda-1bc5-44ab-a081-af6e8740634c"
                }
            }
        }
    }
}